Chane point coordinates type to int

+++ src/chapter15/e15_34/src/main/java/com/dawids/Point.java
@@ -1,31 +1,31 @@
 package com.dawids;
 
 public class Point {
-    private double x = 0.0;
-    private double y = 0.0;
+    private int x = 0;
+    private int y = 0;
 
     public Point() {
 
     }
-   
-    public Point(double x, double y) {
+
+    public Point(int x, int y) {
         this.x = x;
         this.y = y;
     }
 
-    public double getX() {
+    public int getX() {
         return x;
     }
 
-    public void setX(double x) {
+    public void setX(int x) {
         this.x = x;
     }
 
-    public double getY() {
+    public int getY() {
         return y;
     }
 
-    public void setY(double y) {
+    public void setY(int y) {
         this.y = y;
     }
